[MC-314] Reorganize files Create SG_CHAIN_ID_TO_CHAIN_ID mapping for backend
