+++ src/constants/quickswap.ts
@@ -1,0 +1,13 @@
+export const ATK_ADDRESS = '0xf868939ee81f04f463010bc52eab91c0839ef08c';
+export const ICE_ADDRESS = '0xc6C855AD634dCDAd23e64DA71Ba85b8C51E5aD7c';
+export const PFI_ADDRESS = '0xe46B4A950c389e80621d10Dfc398e91613C7e25E';
+export const TREKS_ADDRESS = '0xa65d74f1f047596b2DaFedFdfA327Ccbd499Aa9e';
+export const CROWD_ADDRESS = '0x483dd3425278C1f79F377f1034d9d2CaE55648B6';
+
+export const QUICKSWAP_ONLY_TOKEN_LIST = [
+  ATK_ADDRESS.toLowerCase(),
+  ICE_ADDRESS.toLowerCase(),
+  PFI_ADDRESS.toLowerCase(),
+  TREKS_ADDRESS.toLowerCase(),
+  CROWD_ADDRESS.toLowerCase(),
+];
